Outsource code to a createRulesUtil function
